pyproject.toml
[tool.poetry]
name = "netflix-recommendation-sys"
version = "0.1.0"
description = ""
authors = ["Timileyin <sadiquetimileyin@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
python-dotenv = "^1.0.1"
requests = "^2.31.0"
psycopg2-binary = "^2.9.9"
pandas = "^2.2.0"
pyarrow = "^15.0.0"
pytest-cov = "^4.1.0"
flask = "^3.0.2"
joblib = "^1.3.2"
scikit-learn = "^1.4.1.post1"
matplotlib = "^3.8.3"

[tool.poetry.group.dev.dependencies]
pytest = "8.0.0"
black = "24.1.1"
pylint = "3.0.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
